Fix/optimize sitemap generation (#1828)
* Preheat and cache calls to `get_post_thumbnail_id`

`All_in_One_SEO_Pack_Sitemap::get_images_from_post` calls
`get_the_post_thumbnail_url` which is pretty expensive on large sites
because `get_post_thumbnail_id` does too much for what it is (see
https://core.trac.wordpress.org/ticket/44772).

Since this generates the sitemap for all posts there is no need to fetch
the meta for each post separately, rather fetch all the thumbnail IDs
and cache them statically.

This yields a 2 second improvement (3 seconds down from 5 seconds) with
1000 posts on the latest WordPress version.

* Defer `All_in_One_SEO_Pack_Sitemap::do_sitemaps`

`transition_post_status` is called for every modified post.
When deleting posts in bulk, the
`All_in_One_SEO_Pack_Sitemap::do_sitemaps` is called for every post.

This is not optimal.

A site with 20k posts takes 10 minutes to delete 10 posts in bulk,
because it takes 1 minute to generate the sitemaps for each of the
deleted posts.

Call the `All_in_One_SEO_Pack_Sitemap::do_sitemaps` method on the
`shutdown` action instead, and only once. This has the added benefit of
making sure that the sitemap generation process does not run out of
resources mid-save and does not interfere with callbacks that are called
on the same saving hooks.

* PHP 5.2 and 5.3 compatibility for 393d96

`array_combine` does not like empty arrays, only use when the database
returns results, otherwise leave as empty array.